+++ lib/third.dart
@@ -1,0 +1,225 @@
+//세 번째 페이지
+import 'package:flutter/material.dart';
+
+Widget buildRoundedBox(String text, Color color) {
+  return Container(
+    width: 100,
+    height: 50,
+    decoration: BoxDecoration(
+      borderRadius: BorderRadius.circular(10),
+      color: color,
+    ),
+    child: Center(
+      child: Text(
+        text,
+        style: TextStyle(
+          fontSize: 15,
+          fontWeight: FontWeight.bold,
+          color: Colors.white,
+        ),
+      ),
+    ),
+  );
+}
+
+class ThirdPage extends StatelessWidget {
+  const ThirdPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    // 추가적인 컨테이너 정보
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.orange[400],
+        title: Text(
+          "MonkeyPlace",
+          style: TextStyle(fontSize: 25),
+        ),
+        toolbarHeight: 70,
+        centerTitle: true,
+      ),
+      body: Column(
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 20),
+            child: Row(
+              children: [
+                Align(
+                  alignment: Alignment.centerLeft,
+                  child: ClipOval(
+                    child: Container(
+                      width: 100,
+                      height: 100,
+                      decoration: BoxDecoration(
+                        shape: BoxShape.circle,
+                        image: DecorationImage(
+                            fit: BoxFit.cover,
+                            image: NetworkImage(
+                                "https://kkomin.github.io/dog.jpg")),
+                      ),
+                    ),
+                  ),
+                ),
+                SizedBox(width: 13),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Align(
+                      child: Text(
+                        '이다민',
+                        style: TextStyle(
+                            fontSize: 25, fontWeight: FontWeight.bold),
+                      ),
+                    ),
+                    SizedBox(
+                      height: 8,
+                    ),
+                    Align(
+                      child: Text(
+                        '빅데이터학과 / 컴퓨터공학과',
+                        style: TextStyle(
+                          fontSize: 18,
+                        ),
+                      ),
+                    ),
+                    SizedBox(
+                      height: 10,
+                    ),
+                    Align(
+                      child: Text(
+                        '안녕하세요. 저는 이런 사람입니다.:)',
+                        overflow: TextOverflow.ellipsis,
+                        style: TextStyle(fontSize: 16),
+                        maxLines: 5,
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+          Divider(
+            thickness: 2,
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 5),
+            child: Row(
+              children: [
+                Text(
+                  "키워드 소개",
+                  style: TextStyle(
+                      fontSize: 20,
+                      fontWeight: FontWeight.bold,
+                      color: Colors.black),
+                )
+              ],
+            ),
+          ),
+          // MBTI, 성격, 취미 등
+          Padding(
+            padding: const EdgeInsets.all(10),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                buildRoundedBox('MBTI', Colors.blue),
+                SizedBox(width: 10),
+                buildRoundedBox('MBTI', Colors.blue),
+                SizedBox(width: 10),
+                buildRoundedBox('MBTI', Colors.blue),
+              ],
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.all(10),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                buildRoundedBox('MBTI', Colors.blue),
+                SizedBox(width: 10),
+                buildRoundedBox('MBTI', Colors.blue),
+                SizedBox(width: 10),
+                buildRoundedBox('MBTI', Colors.blue),
+              ],
+            ),
+          ),
+          SizedBox(
+            height: 5,
+          ),
+          Divider(
+            thickness: 2,
+          ),
+          Row(
+            children: [
+              Padding(
+                padding:
+                    const EdgeInsets.symmetric(horizontal: 30, vertical: 5),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      "자기소개",
+                      style: TextStyle(
+                          fontSize: 20,
+                          fontWeight: FontWeight.bold,
+                          color: Colors.black),
+                    ),
+                    SizedBox(
+                      height: 10,
+                    ),
+                    Align(
+                      alignment: Alignment.topLeft,
+                      child: Text(
+                        "자기소개 하기~ 얍!",
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 10,
+                        style: TextStyle(fontSize: 17),
+                      ),
+                    ),
+                    SizedBox(
+                      height: 70,
+                    )
+                  ],
+                ),
+              ),
+            ],
+          ),
+          Divider(
+            thickness: 2,
+          ),
+          Row(
+            children: [
+              Padding(
+                padding:
+                    const EdgeInsets.symmetric(horizontal: 30, vertical: 5),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      "자격증",
+                      style: TextStyle(
+                          fontSize: 20,
+                          fontWeight: FontWeight.bold,
+                          color: Colors.black),
+                    ),
+                    SizedBox(
+                      height: 10,
+                    ),
+                    Align(
+                      alignment: Alignment.topLeft,
+                      child: Text(
+                        "oooo\noooooo",
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 10,
+                        style: TextStyle(fontSize: 17),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
